backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
qdrant-client==1.16.1
openai==2.14.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.9.2
Markdown==3.10
